Move the classes extending PrefsListItemWrapper into PrefsListItemWrapper.kt and rename it to PrefsListItemWrapperClasses.kt for clarity.
